requirements.txt
networkx==2.8.8
pydot==1.4.1
scipy==1.9.3
pyparsing==3.0.1
flask==2.2.2
flask_cors==3.0.10
python-dotenv==0.21.0
numpy==1.24.1